+++ contrib/grpcplugins/action/examples/nodejs/package.json
@@ -16,5 +16,9 @@
     "get-port": "^4.0.0",
     "google-protobuf": "^3.6.1",
     "grpc": "^1.15.1"
+  },
+  "devDependencies": {
+    "webpack": "^4.12.0",
+    "webpack-cli": "^3.1.2"
   }
 }
